requirements.txt
asyncua==1.0.0
pyyaml==6.0.1
cryptography>=41.0.0
pandas>=1.5.0